python-service/requirements.txt
# Python Service Requirements
# Book PDF/DOCX to Structured Content Converter

# ============================================
# WEB FRAMEWORK
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ============================================
# PDF PROCESSING
# ============================================
PyMuPDF==1.23.8
pdfplumber==0.10.3
pdf2image==1.17.0

# ============================================
# DOCX PROCESSING
# ============================================
python-docx==1.1.0

# ============================================
# OCR (Optical Character Recognition)
# ============================================
pytesseract==0.3.10
easyocr==1.7.1
Pillow==10.2.0

# ============================================
# ARABIC TEXT PROCESSING
# ============================================
arabic-reshaper==3.0.0
python-bidi==0.4.2

# ============================================
# AI - LOCAL (Free)
# ============================================
ollama==0.3.3
langchain==0.1.0
langchain-community==0.0.13

# ============================================
# AI - PAID (Optional)
# ============================================
anthropic==0.18.1
openai==1.12.0
google-cloud-vision==3.5.0

# ============================================
# DATABASE
# ============================================
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# ============================================
# UTILITIES
# ============================================
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
aiohttp==3.9.1

# ============================================
# TESTING
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
